unit test: controllers

+++ src/main/java/br/com/testeintegracao/controller/BookController.java
@@ -1,8 +1,11 @@
 package br.com.testeintegracao.controller;
 
+import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,6 +14,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import br.com.testeintegracao.model.Book;
 import br.com.testeintegracao.service.BookService;
@@ -27,19 +31,43 @@ public class BookController {
 		return bookService.getBooks();
 	}
 	
+	@GetMapping("/{id}")
+	public ResponseEntity<Book> getBook(@PathVariable int id) {
+		Optional<Book> book = bookService.getBookById(id);
+		if(book.isPresent()) {
+			return ResponseEntity.ok().body(book.get());
+		}
+		return ResponseEntity.noContent().build();
+	}
+	
 	@PostMapping("/create")
-	public Book createBook(@RequestBody Book book) {
-		return bookService.create(book);
+	public ResponseEntity<Book> createBook(@RequestBody Book book, UriComponentsBuilder uriBuilder) {		
+		Optional<Book> newBook = bookService.create(book);		
+		if(newBook.isPresent()) {
+			URI uri = uriBuilder.path("/book/{id}").buildAndExpand(newBook.get().getId()).toUri();
+			return ResponseEntity.created(uri).body(newBook.get());			
+		}
+		return ResponseEntity.badRequest().build();		
 	}
 	
 	@PutMapping("/put")
-	public Book modifyBook(@RequestBody Book book) {
-		return bookService.modify(book);
+	public ResponseEntity<Book> modifyBook(@RequestBody Book updatedBook) {
+		Optional<Book> bookToUpdate = bookService.getBookById(updatedBook.getId());
+        if(bookToUpdate.isPresent()) {
+        	Book modifiedBook = bookService.modify(updatedBook);
+        	return ResponseEntity.ok().body(modifiedBook);
+        }
+        return ResponseEntity.noContent().build();
 	}
 	
 	@DeleteMapping("/delete/{id}")
-	public void deleteBook(@PathVariable("id") int id) {
-		bookService.delete(id);
+	public ResponseEntity<Book> deleteBook(@PathVariable("id") int id) {
+		Optional<Book> bookToDelete = bookService.getBookById(id);
+		if(bookToDelete.isPresent()) {
+			bookService.delete(id);
+			return ResponseEntity.ok().build();
+		}
+        return ResponseEntity.noContent().build();
 	}
 
 }
